Origin/feature/branch ftom feature

+++ src/main/kotlin/com/kardibus/temp/dto/ModelDto.kt
@@ -2,8 +2,8 @@ package com.kardibus.temp.dto
 
 class ModelDto {
     var id: Long? = null
-    var prog: Int = 0
-    var curr: Int = 0
+    var prog: Long = 0
+    var curr: Long = 0
     var temp: Double = 0.0
     var work: Boolean = false
 }
