+++ src/main/kotlin/com/kardibus/temp/model/Model.kt
@@ -9,8 +9,8 @@ class Model {
     @Id
     @GeneratedValue
     var id: Long? = null
-    var prog: Int = 0
-    var curr: Int = 0
+    var prog: Long = 0
+    var curr: Long = 0
     var temp: Double = 0.0
     var work: Boolean = false
     override fun toString(): String {
